http: fix wrong variable scope for asynchronous functions
When calling the POST and PUT asynchronous functions, the "body"
parameter was passed as a reference, which would lead to not being
taken into account if its scope was not maintained by the caller
application. This patch fixes the issue by copying the body string
before calling the asynchronous part of the PUT/POST functions.

Change-Id: I1fd97a21ecdccf015c9e83d56eb9752cd3f19d0b
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>