security: return singleton when calling request/release
The artiksee OpenSSL engine cannot be initialized twice within the
same process, which happens if the application runs two SDR based
services (e.g. LWM2M and websocket) in the same process. This patch
modifies the request/release pair of the security module to always
return a reference to the same instance of the security object, while
counting references for destroying the object only when  all references
have been released.

Change-Id: Ie43124024c8ab62bf612745f2e3a7a544acef791
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>