linux/spi: fix wrong transaction structure initalization
When calling the spidev driver from the SPI module, there are
two places where the max frequency and word size should be set.
The first one was done at request time using associated ioctls (1),
and the other one when filling up the transaction structure (2).

In our case we would do (1) but just leave 0 for these parameters
when doing (2). It seems some newer version of the Linux kernel does
not behave correctly in this case, leading to reading wrong values out
of the SPI bus.

This patch fixes the issue by setting the parameters during steps
(1) and (2), which makes the SPI module behave correctly with all
versions of the kernel.

Change-Id: Ib980c22bd5885a34e9e051bbff37aa40229eb13e
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>