tizenrt/http: return HTTP status code even on failure
When the server returns a status code different than 200, the
HTTP request functions return an error code but do not return the
status code. Calling application knows that there was an issue but
they don't get the associated status code from the server, which is
often a good indication of the failure. This patch improves the returned
information by filling up the status output parameter even in the
case where an error code is returned.

Change-Id: I716de7c692c3a21288c0848e05af0b3ab4ae5837
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>