connectivity/network: improve network monitoring API
This patch modifies the 'get_online_status' API to be
more flexible by taking as parameters the remote server
to ping, the periodicity of the pings, the timeout after
which the connectivity to the remote server is considered
to be down.

This is especially useful to accomodate regional differences
and connectivity conditions across various sites where a
single application could be running.

Change-Id: I7cf14a411b3c166a29364ff2e2b7032033e19b93
Signed-off-by: Camille Bégué <c.begue@samsung.com>