websocket: send event disconnect when connectivity goes down
This patch adds handling for the case where connectivity gets down
due to bad or faulty external connection. This kind of event was
previously not being detected and the websocket would remain open
without notifying the calling application.

This patch adds detection of such condition and sends an event
that will trigger the appropriate 'connection' callback with
proper error code passed to the registered application.

The detection is performed by implementing the ping/pong mechanism
detailed in RFC-6455. The websocket configuration structure exposes
new fields to fine-tune timeout and periodicity of the ping/pong
mechanism to let the application accomodate specific use cases.
Setting a 'ping_period' parameter of 0 disables the ping/pong
monitoring.

Change-Id: Ibeed42d28460e50fdc42d1104e17d3149499d6f1
Signed-off-by: Geoffroy Letourneur <g.letourneur@samsung.com>
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>