linux/serial: fix blocked port after wrongfully setting hard flow control
When we open a serial port with hardware flow control set, if the
RTS/CTS wires are not routed then the port blocks. When trying to reopen
this same port without setting flow control, it would still fail to open
and block. This was due to the API reusing the current termios
configuration that happens to have an undocumented bit set (bit 31).
Clearing this bit reverts to normal behavior. This patch fixes this
issue by starting with a clean termios config every time the port is
opened.

Change-Id: I370b0e392b7175e08c2a96a6a55e561dba878ae7
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>