tizenrt/http: fix memory leak in asynchronous functions
This patch fixes a memory leak that occurs when calling the
asynchronous HTTP functions. Previously the URL of the HTTP
request was duplicated but associated memory was not released
upon success or failure, thus leaking the size of the URL.

This patch fixes the issue by adding proper release calls
where expected.

Change-Id: I3659aa2a52c9322f42d74163b2a658cc25ff5d4a
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>