test/i2c: add automatic rebinding of the driver after testing
The I2C test unbinds a driver linked to a I2C chip prior to
running a test sending transactions to that chip. This is needed
since the driver keeps the associated I2C bus/chip busy, thus the
application cannot request it to perform the tests.

This patch adds a final step at the end of the test to rebind the
driver so that it can operate properly after the test is over.

Change-Id: Ib99867134635c334cc542820971e9aa57604fbc6
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>