tizenrt/http: fix pointer double-free
The 'init_client_ssl_config" function releases allocated memory
in case something goes wrong in the process. However it used to
return an allocated pointer as an out parameter even if the function
fails. As a result the calling function saw this pointer as valid,
even though it was freed by the function before returning, and was
trying to release it again.

This patch ensures the pointer is returned by 'init_client_ssl_config'
only if the function was successful.

Change-Id: Ic8ffbd0d7e9fab9fe46e772e3cb0b50e11668156
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>