linux/wifi: fix handling of mixed ASCII/unicode SSIDs
When checking for unicode SSIDs, only the first byte of the
string would be checked, which caused mixed strings starting
with ASCII followed by unicode to not be properly recognized
as containing unicode and would be wrongfully processed as
ASCII only strings.

This patch fixes the issue by checking for the '\' character
in the whole string in order to determine if it should be
processed as a unicode string or a plain ASCII one.

Change-Id: I8c9f5797747497024beed275d0dd2e31cc07c754
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>