linux/mqtt: fix handling of keepalive mechanism
Previously, the mosquitto's "misc" loop function was only called
whenever some activity was detected on the MQTT socket. As a result
the pings were not handled periodically by mosquitto, which would
result in the keepalive period to be reached without any ping being
handled, and the MQTT connection would be closed.

This patch adds a periodic callback to be executed every half
keepalive period, which calls mosquitto's "misc" loop to handle
the ping requests/responses as expected. The test programs have
also been modified to add a reasonable 5 seconds keepalive period.

Change-Id: Iae1f78797a697f9519f55ae65fd7dcda1b211753
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>