tizenrt/adc: fix issue when doing multiple reads in a row
Previously the open call to the ADC driver was done in the
'request' function and the fd was closed after reading in the
'read' function, which caused subsequent reads to fail since
the fd was closed.

This patch fixes the issue by doing open/close on the driver
once in the 'request' function to check the pin number parameter,
then calls open/close again for each read of the ADC value. It
seems the latter is needed to properly reinitialize the ADC
conversion between two reads, otherwise the conversion sometimes
returns irrelevant values on the second read.

Change-Id: I2a63436e53d53c2928ceca42043fe5d65ad71592
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>