packaging: remove dependency to artiksee
This patch removes the runtime dependency to the artiksee
OpenSSL engine. The main reason is to avoid circular
dependency since artiksee also depends on the ARTIK SDK.

This removal is considered acceptable since only few APIs
of the ARTIK SDK require the OpenSSL engine, generally
when secure connection to ARTIK Cloud is requested. Failing
to install the package and calling these APIs anyway will
result in an explicit enough error that the developer should
easily understand and fix.

Change-Id: I43c439124240938b15ce85bbaebd73fe215dc40b
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>