tizenrt/coap: fix DTLS timeout error and timeout inactive session
This patch fixes a timeout error that was previously logged
even in the event of a successful conncection. It also increases
the timeout after which the server will kill an inactive client
session. It was previously set to 30 seconds and is now defined
to be 5 minutes so that clients can keep connected to the server
for a longer time without the need to disconnect/reconnect.

Change-Id: I0333de914c368cb80605790face77b4eff24d44f
Signed-off-by: Romain Cancilliere <romain.c@samsung.com>