linux/time: fix issue in computing delta when creating alarm
This patch fixes an issue that would prevent proper computation
of the delta number of seconds when creating an alarm. The issue
lied in the fact the code assumed an 'artik_time_date' structure
could be mapped on top of a 'tm' structure, which is not strictly
the case. This patch fixes the problem by copying date fields one
by one from one structure to the other.

Change-Id: Ia750d7a26fb875f99af94eaa5172a0af92b736d0
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>