linux/websocket: fix segfault when connecting to an invalid URI
This patch fixes a segfault in the linux websocket part.
The 'lws_connection_via_info' function from libwesockets was called
before the 'conn_info' structure is initialized. As a result it may
be accessed from the connection error callback before it is actually
initialized. This patchs solves the issue by changing the initialization
order and moves the URI parsing at 'request' time.

Change-Id: I4cc8a7210a316181ee7aba89286441c99aee0ec7
Signed-off-by: Geoffroy Letourneur <g.letourneur@samsung.com>