linux/websocket: fix parsing root CA certificate
When we give a PEM with a chain of root CA certificates,
we parse it in order to extract the different certificates
and add them to the keystore. The problem was the incrementation
of the size after finding the "END CERTIFICATE" header and
the pointer didn't contain the beginning of the certificate.

Change-Id: Ice7ccf2e14b84c27c997b1a1c91bb8f473a394ce
Signed-off-by: Romain Cancilliere <romain.c@samsung.com>