websocket: fix URI parsing in artik common layer
A previous patch moves parsing the URI to the OS independent
layer but it was not properly applied to the OS layers for
both TizenRT and Linux. This patch fixes this by having the
common layer do the parsing at request time to reject potentially
malformed URIs, then store the extracted information to pass it
to the OS layer when needed at connection time.

It makes the whole code cleaner, more reusable, and removes some
memory allocations that were only there because of the multiple
layers.

Change-Id: I52c0b43fab1a3c648dc946041ccfb41a35608ca9
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>