tests: specify algorithm when calling 'get_publickey'
Non-S module only have ARTIK certificate and key in SE
based on ECC_SEC_P256R1 alogorithm. When using this
function to get the public key prior to initiating a
SDR connection to the server, it fails as it cannot
properly find the public key from the Secure Element.

This patch fixes this issue by explicitely setting the
algorithm in test applications to work with non-S
modules.

Change-Id: Ib1c97c8265adc3ac338f1c6a29773393a4cd0f5b
Signed-off-by: xiangfei.ma <xiangfei.ma@samsung.com>
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>