tizenrt/coap: do not call DTLS receive function after new session
Timeout from server-side was removed in a previous patch, that
causes a blocking at DTLS receive function after having
created new DTLS session (due to select() function).

This patch introduces a new parameter that verifies if a DTLS
session is newly created, in which case the server waits until
it receives a first packet before calling the blocking 'select()'.

Change-Id: I6cfffe9d4e2067b645bc442a976fa54809b1d5b0
Signed-off-by: Romain Cancilliere <romain.c@samsung.com>